requirements.txt
apache-airflow
requests
azure-cli
apache-airflow[celery]
pyspark
python-dotenv
azure-storage-file-datalake
pyarrow
streamlit
pywhatkit
mlflow